Comparers rename
