+++ src/Nethermind/Nethermind.Consensus.AuRa/Transactions/CompareTxByPriorityBase.cs
@@ -28,14 +28,14 @@ using Nethermind.TxPool;
 
 namespace Nethermind.Consensus.AuRa.Transactions
 {
-    public abstract class CompareTxByPermissionBase : IComparer<Transaction>
+    public abstract class CompareTxByPriorityBase : IComparer<Transaction>
     {
         private readonly IContractDataStore<Address> _sendersWhitelist;
         private readonly IDictionaryContractDataStore<TxPriorityContract.Destination> _priorities;
         private Keccak _blockHash;
         private ISet<Address> _sendersWhiteListSet;
 
-        public CompareTxByPermissionBase(IContractDataStore<Address> sendersWhitelist, // expected HashSet based
+        public CompareTxByPriorityBase(IContractDataStore<Address> sendersWhitelist, // expected HashSet based
             IDictionaryContractDataStore<TxPriorityContract.Destination> priorities) // expected SortedList based)
         {
             _sendersWhitelist = sendersWhitelist ?? throw new ArgumentNullException(nameof(sendersWhitelist));
